features/shard: Do list_del_init() while list memory is valid
Problem:
shard_post_lookup_fsync_handler() goes over the list of inode-ctx that need to
be fsynced and in cbk it removes each of the inode-ctx from the list. When the
first member of list is removed it tries to modifies list head's memory with
the latest next/prev and when this happens, there is no guarantee that the
list-head which is from stack memory of shard_post_lookup_fsync_handler() is
valid.

Fix:
Do list_del_init() in the loop before winding fsync.

 >BUG: 1557876
 >Change-Id: If429d3634219e1a435bd0da0ed985c646c59c2ca
 >Signed-off-by: Pranith Kumar K <pkarampu@redhat.com>

Upstream-patch: https://review.gluster.org/19737
BUG: 1556895
Change-Id: If429d3634219e1a435bd0da0ed985c646c59c2ca
Signed-off-by: Pranith Kumar K <pkarampu@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/133241
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>