shared storage: Prevent mounting shared storage from non-trusted client
gluster shared storage is a volume used for internal storage for
various features including ganesha, geo-rep, snapshot.

So this volume should not be exposed to the client, as it is
a special volume for internal use.

This fix wont't generate non trusted volfile for shared storage volume.

backport of https://review.gluster.org/#/c/19920/

>Change-Id: I8ffe30ae99ec05196d75466210b84db311611a4c
>updates: bz#1570432
>Signed-off-by: Mohammed Rafi KC <rkavunga@redhat.com>

Change-Id: Ic540b983bcc53a783fda7ca7a283a9ab48d9eeb7
BUG: 1568969
Signed-off-by: Mohammed Rafi KC <rkavunga@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/136708
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Atin Mukherjee <amukherj@redhat.com>