cluster/dht: Handle file migrations when brick down
The decision as to which node would migrate a file
was based on the gfid of the file. Files were divided
among the nodes for the replica/disperse set. However,
if a brick was down when rebalance started, the nodeuuids
would be saved as NULL and a set of files would not be migrated.

Now, if the nodeuuid is NULL, the first non-null entry in
the set is the node responsible for migrating the file.

upstream master: https://review.gluster.org/#/c/19831/

> Change-Id: I72554c107792c7d534e0f25640654b6f8417d373
> fixes: bz#1564198
> Signed-off-by: N Balachandran <nbalacha@redhat.com>

Change-Id: Ia0e15339aefee2712e85d7e282c9b7934665376b
BUG: 1553677
Signed-off-by: N Balachandran <nbalacha@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/135515
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Atin Mukherjee <amukherj@redhat.com>