Core: The lock contention on gf_client_dump_inodes_to_dict
        backport of:https://review.gluster.org/#/c/20035/

Problem: For a distributed replicated volume, in the inode status
command the lock on gf_client_dump_inodes_to_dict is held by the
first brickop. while this is being processed, if the second brickop
comes. It fails to get the lock and the whole brick op fails.

Fix: Instead of using a TRY_LOCK which errors out if the lock is busy,
Use LOCK which will wait till the lock is acquired.

>Change-Id: I768a0a1b60f28c4f7f94549e18ee3765b69cc528
>Signed-off-by: hari gowtham <hgowtham@redhat.com>
>BUG: 1559452

fixes: bz#1579769
Change-Id: I768a0a1b60f28c4f7f94549e18ee3765b69cc528
BUG: 1559452
Signed-off-by: hari gowtham <hgowtham@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/140334
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>