glusterd: show brick online after port registration even in brick-mux
	Upstream-patch: https://review.gluster.org/20451

Problem:
With brick-mux even before brick attach is complete on the bricks
glusterd marks them as online. This can lead to a race where
scripts that check if the bricks are online to assume that the
brick is online before it is completely online.

Fix:
Wait for the callback from the brick before marking the port
as registered so that volume status will show the correct status
of the brick.

 >fixes bz#1597568
BUG: 1598356
Change-Id: Icd3dc62506af0cf75195e96746695db823312051
Signed-off-by: Pranith Kumar K <pkarampu@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/143592
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>