posix: Do not log ENXIO errors for seek fop
When lseek is used with SEEK_DATA and SEEK_HOLE, it's expected that the
last operation fails with ENXIO when offset is beyond the end of file.
In this case it doesn't make sense to report this as an error log message.

This patch reports ENXIO failure messages for seek fops in debug level
instead of error level.

> Change-Id: I62a4f61f99b0e4d7ea6a2cdcd40afe15072794ac
> fixes: bz#1598926
> Signed-off-by: Xavi Hernandez <xhernandez@redhat.com>

Upstream patch: https://review.gluster.org/c/glusterfs/+/20475
Change-Id: I62a4f61f99b0e4d7ea6a2cdcd40afe15072794ac
BUG: 1598883
Signed-off-by: Xavi Hernandez <xhernandez@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/158531
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>