glusterd: memory leak in get-state
Problem: gluster get-state command is leaking the memory when
geo-replication session is configured.

Cause: In glusterd_print_gsync_status(), we are trying to get
reference to the keys of gsync_dict. The references to keys of
gsync_dict are stored status_vols[i]. status_vols[i] are
allocated with a memory of size of gf_gsync_status_t.

Solution: Need not to use a array of pointers(status_vals), using
a pointer to hold the reference to a key of gsync_dict is sufficient.

Followed the below steps for testing:
1. Configured geo-rep session
2. Ran gluster get-state command for 1000 times.

Without this patch, glusterd's memory was increasing significantly
(around 22000KB per 1000 times), with this patch it reduced (1500KB
per 1000 times)

>fixes: bz#1601423
>Change-Id: I361f5525d71f821bb345419ccfdc20ca288ca292
>Signed-off-by: Sanju Rakonde <srakonde@redhat.com>

upstream patch: https://review.gluster.org/#/c/20521/

Change-Id: I361f5525d71f821bb345419ccfdc20ca288ca292
BUG: 1599362
Signed-off-by: Sanju Rakonde <srakonde@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/144325
Tested-by: RHGS Build Bot <nigelb@redhat.com>
Reviewed-by: Mohit Agrawal <moagrawa@redhat.com>
Reviewed-by: Sunil Kumar Heggodu Gopala Acharya <sheggodu@redhat.com>