posix: remove not supported get/set content
getting and setting a file's content using extended
attribute worked great as a GET/PUT alternative when
an object storage is supported on top of Gluster. But
it needs application changes, and also, it skips some
caching layers. It is not used over years, and not
supported any more. Remove the dead code.

BUG: 1608363
BUG: 1608366
Change-Id: I6d4c135ca1bce5197c7cf749f9f3b481b91f2a95
Signed-off-by: Amar Tumballi <amarts@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/145896
Reviewed-by: Pranith Kumar Karampuri <pkarampu@redhat.com>
Reviewed-by: FNU Raghavendra Manjunath <raghavendra@redhat.com>
Reviewed-by: Atin Mukherjee <amukherj@redhat.com>