glusterd: more stricter checks of if brick is running in multiplex mode
While gf_attach () utility can help in detaching a brick instance from
the brick process which the kill_brick () function in tests/volume.rc
uses it has a caveat which is as follows:
1. It doesn't ensure the respective brick is marked as stopped which
glusterd does from glusterd_brick_stop
2. Sometimes if kill_brick () is executed just after a brick stack is
up, the mgmt_rpc_notify () can take some time before marking
priv->connected to 1 and before it if kill_brick () is executed, brick
will fail to initiate the pmap_signout which would inturn cleans up the
pidfile.

To avoid such possibilities, a more stricter check on if a brick is
running or not in brick multiplexing has been brought in now where it
not only checks for its pid's existance but checks if the respective
process has the brick instance associated with it before checking for
brick's status.

> Change-Id: I98b92df949076663b9686add7aab4ec2f24ad5ab
> Signed-off-by: Atin Mukherjee <amukherj@redhat.com>
> (cherry pick from commit bb9b8f61501cc633e585593de4d9f2fe5494d5ce)
> (Reviewed on upstream link https://review.gluster.org/#/c/glusterfs/+/20651/)

Change-Id: I47b70927c839ca4828a0499006b5c1f604d3d6a4
BUG: 1612098
Signed-off-by: Mohit Agrawal <moagrawa@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/146874
Reviewed-by: Atin Mukherjee <amukherj@redhat.com>
Tested-by: Atin Mukherjee <amukherj@redhat.com>