io-stats: sanitize the dump path further
In the previous patch, while addressing the comment on review,
a "/" at the end of the "/var/run/gluster" directory was missed out.

Also noticed that the logic to convert the '/' to '-' for sanity
of the path needed to change.

Testing: Ran the tests which marked the bug as FailedQA,
and also validated the originally reported issue, and now we see
a specific log when not so clean path is given as the value to
this xattr.

BUG: 1605086
Change-Id: Ia8397ecd5841a72d0daca0106557e1226c293e35
Signed-off-by: Amar Tumballi <amarts@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/147644
Reviewed-by: Vijay Bellur <vbellur@redhat.com>
Reviewed-by: Atin Mukherjee <amukherj@redhat.com>
Tested-by: Atin Mukherjee <amukherj@redhat.com>